Move `CommandStorage` under the `command` package
