+++ server/src/main/java/org/spine3/server/storage/memory/InMemoryCommandStorage.java
@@ -26,7 +26,7 @@ import org.spine3.base.CommandId;
 import org.spine3.base.CommandStatus;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
-import org.spine3.server.storage.CommandStorage;
+import org.spine3.server.command.CommandStorage;
 import org.spine3.server.storage.CommandStorageRecord;
 
 import javax.annotation.Nullable;
