+++ server/src/test/java/org/spine3/server/command/CommandStorageShould.java
@@ -18,9 +18,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.storage;
+package org.spine3.server.command;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
 import com.google.protobuf.Any;
 import com.google.protobuf.StringValue;
 import org.junit.After;
@@ -34,6 +35,8 @@ import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
 import org.spine3.protobuf.AnyPacker;
+import org.spine3.server.storage.AbstractStorageShould;
+import org.spine3.server.storage.CommandStorageRecord;
 import org.spine3.test.Tests;
 import org.spine3.test.storage.ProjectId;
 import org.spine3.test.storage.command.CreateProject;
@@ -93,7 +96,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Override
     protected CommandStorageRecord newStorageRecord() {
-        final Any command = AnyPacker.pack(Given.Command.createProject());
+        final Any command = AnyPacker.pack(org.spine3.server.storage.Given.Command.createProject());
         final String commandType = ofEnclosed(command).getTypeName();
         final CommandContext context = createCommandContext();
         final CommandStorageRecord.Builder builder = CommandStorageRecord.newBuilder()
@@ -119,7 +122,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void store_and_read_command() {
-        final Command command = Given.Command.createProject();
+        final Command command = org.spine3.server.storage.Given.Command.createProject();
         final CommandId commandId = getId(command);
 
         storage.store(command);
@@ -130,7 +133,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void store_command_with_error() {
-        final Command command = Given.Command.createProject();
+        final Command command = org.spine3.server.storage.Given.Command.createProject();
         final CommandId commandId = getId(command);
         final Error error = newError();
 
@@ -143,11 +146,11 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void store_command_with_error_and_generate_ID_if_needed() {
-        final Command command = Commands.create(Given.CommandMessage.createProject(), CommandContext.getDefaultInstance());
+        final Command command = Commands.create(org.spine3.server.storage.Given.CommandMessage.createProject(), CommandContext.getDefaultInstance());
         final Error error = newError();
 
         storage.store(command, error);
-        final List<CommandStorageRecord> records = newArrayList(storage.read(ERROR));
+        final List<CommandStorageRecord> records = Lists.newArrayList(storage.read(ERROR));
 
         assertEquals(1, records.size());
         assertFalse(records.get(0)
@@ -158,7 +161,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void store_command_with_status() {
-        final Command command = Given.Command.createProject();
+        final Command command = org.spine3.server.storage.Given.Command.createProject();
         final CommandId commandId = getId(command);
         final CommandStatus status = SCHEDULED;
 
@@ -170,12 +173,12 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void load_commands_by_status() {
-        final List<Command> commands = ImmutableList.of(Given.Command.createProject(), Given.Command.addTask(), Given.Command.startProject());
+        final List<Command> commands = ImmutableList.of(org.spine3.server.storage.Given.Command.createProject(), org.spine3.server.storage.Given.Command.addTask(), org.spine3.server.storage.Given.Command.startProject());
         final CommandStatus status = SCHEDULED;
 
         store(commands, status);
         // store an extra command with another status
-        storage.store(Given.Command.createProject(), ERROR);
+        storage.store(org.spine3.server.storage.Given.Command.createProject(), ERROR);
 
         final Iterator<Command> iterator = storage.iterator(status);
         final List<Command> actualCommands = newArrayList(iterator);
@@ -229,7 +232,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     @Test
     public void convert_cmd_to_record() {
-        final Command command = Given.Command.createProject();
+        final Command command = org.spine3.server.storage.Given.Command.createProject();
         final CommandStatus status = RECEIVED;
 
         final CommandStorageRecord record = CommandStorage.newRecordBuilder(command, status).build();
@@ -296,19 +299,19 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
     @Test(expected = IllegalStateException.class)
     public void throw_exception_if_try_to_store_cmd_to_closed_storage() {
         close(storage);
-        storage.store(Given.Command.createProject());
+        storage.store(org.spine3.server.storage.Given.Command.createProject());
     }
 
     @Test(expected = IllegalStateException.class)
     public void throw_exception_if_try_to_store_cmd_with_error_to_closed_storage() {
         close(storage);
-        storage.store(Given.Command.createProject(), newError());
+        storage.store(org.spine3.server.storage.Given.Command.createProject(), newError());
     }
 
     @Test(expected = IllegalStateException.class)
     public void throw_exception_if_try_to_store_cmd_with_status_to_closed_storage() {
         close(storage);
-        storage.store(Given.Command.createProject(), OK);
+        storage.store(org.spine3.server.storage.Given.Command.createProject(), OK);
     }
 
     @Test(expected = IllegalStateException.class)
@@ -362,7 +365,7 @@ public abstract class CommandStorageShould extends AbstractStorageShould<Command
 
     private static Failure newFailure() {
         return Failure.newBuilder()
-                      .setInstance(Given.EventMessage.projectCreatedAny())
+                      .setInstance(org.spine3.server.storage.Given.EventMessage.projectCreatedAny())
                       .setStacktrace("failure stacktrace")
                       .setTimestamp(getCurrentTime())
                       .build();
