+++ server/src/test/java/org/spine3/server/storage/memory/InMemoryCommandStorageShould.java
@@ -20,8 +20,8 @@
 
 package org.spine3.server.storage.memory;
 
-import org.spine3.server.storage.CommandStorage;
-import org.spine3.server.storage.CommandStorageShould;
+import org.spine3.server.command.CommandStorage;
+import org.spine3.server.command.CommandStorageShould;
 
 public class InMemoryCommandStorageShould extends CommandStorageShould {
